Puts the trace panels' layouts and plotting functions into individual files
