+++ NuRadioReco/eventbrowser/apps/trace_plots/rec_electric_field_spectrum.py
@@ -4,8 +4,25 @@ import plotly
 from NuRadioReco.utilities import units
 from NuRadioReco.eventbrowser.default_layout import default_layout
 import numpy as np
+import dash_html_components as html
+import dash_core_components as dcc
+from dash.dependencies import Input, Output, State
+from app import app
+import dataprovider
+provider = dataprovider.DataProvider()
 
-def update_efield_spectrum(trigger, evt_counter, filename, station_id, juser_id, provider):
+layout = [
+dcc.Graph(id='efield-spectrum')
+]
+
+@app.callback(
+    dash.dependencies.Output('efield-spectrum', 'figure'),
+    [dash.dependencies.Input('trigger-trace', 'children'),
+     dash.dependencies.Input('event-counter-slider', 'value'),
+     dash.dependencies.Input('filename', 'value'),
+     dash.dependencies.Input('station-id-dropdown', 'value')],
+     [State('user_id', 'children')])
+def update_efield_spectrum(trigger, evt_counter, filename, station_id, juser_id):
     if filename is None or station_id is None:
         return {}
     user_id = json.loads(juser_id)
